requirements.txt
about-time==4.2.1
aiohttp==3.8.5
aiosignal==1.3.1
alive-progress==3.1.4
annotated-types==0.5.0
anyio==3.7.1
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bce-python-sdk==0.8.87
beautifulsoup4==4.12.3
bidict==0.22.1
bleach==6.1.0
blinker==1.6.2
Bottleneck==1.3.5
cachetools==5.3.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
cirq-core==1.1.0
cirq-google==1.1.0
click==8.1.6
colorama==0.4.6
colorlog==6.7.0
comm==0.1.4
contourpy==1.0.5
cvxpy==1.3.2
cycler==0.11.0
data==0.4
datasets==2.14.4
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.4
docopt==0.6.2
duet==0.2.8
ecos==2.0.12
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.101.0
fastdtw==0.3.4
fastjsonschema==2.19.1
filelock==3.12.2
Flask==2.3.2
Flask-Babel==2.0.0
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2023.6.0
funcsigs==1.0.2
future==0.18.3
google-api-core==1.34.0
google-auth==2.22.0
googleapis-common-protos==1.60.0
grapheme==0.6.0
grpcio==1.57.0
grpcio-status==1.48.2
h11==0.14.0
h5py==3.9.0
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.0
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.4
latex==0.7.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2

mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.12.2
munkres==1.1.4
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==2.8.8
numexpr==2.8.4
numpy==1.23.4
opencv-python==4.8.0.76
openfermion==1.5.1
openpyxl==3.0.10
opt-einsum==3.3.0
osqp==0.6.3
packaging==23.1
paddle-bfloat==0.1.2
paddle-quantum==2.4.0
paddle2onnx==1.0.6
paddlefsl==1.1.0
paddlenlp==2.5.2
paddlepaddle==2.3.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
PubChemPy==1.0.4
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.18.0
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
pyparsing==3.0.9
pyprimes==0.1

python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
qcompute==2.0.4
qdldl==0.1.7.post0
referencing==0.33.0
reportlab==4.0.9
requests==2.31.0
rich==13.5.2
rpds-py==0.18.0
rsa==4.9
SciencePlots==2.1.0
scikit-learn==1.3.0
scipy==1.10.1
scs==3.2.3
seaborn==0.12.2
sentencepiece==0.1.99
seqeval==1.2.2

setuptools==66.1.1

shutilwhich==1.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tempdir==0.7.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
visualdl==2.4.2
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.6
xxhash==3.3.0
yarg==0.1.9
yarl==1.9.2
zipp==3.11.0
